release: Chromium 32.0.1700 (32.0.1700.99)
Compare: https://github.com/chromium/chromium/compare/3a6a9a34b4b3eea04c3ef652ba031dfec2f89615...539909f5a9b54507d4b735e1346addd9e8bfbcf1

+++ media/cast/rtp_sender/rtp_packetizer/rtp_packetizer.h
@@ -45,7 +45,8 @@ class RtpPacketizer {
   uint16 NextSequenceNumber();
 
   int send_packets_count() { return send_packets_count_; }
-  int send_octet_count() { return send_octet_count_; }
+
+  size_t send_octet_count() { return send_octet_count_; }
 
  private:
   void Cast(bool is_key, uint8 reference_frame_id,
@@ -64,7 +65,7 @@ class RtpPacketizer {
   uint16 packet_id_;
 
   int send_packets_count_;
-  int send_octet_count_;
+  size_t send_octet_count_;
 };
 
 }  // namespace cast
